Updated Particles Massively, and the Color Scheme of the Website
The color scheme of the website was updated once again, hopefully for the last time. The themes for the light mode and dark mode were both fleshed out.

The Particles have been updated from their default state to something subtler.

Work was done to have the particles update color when switching between light and dark mode.